Time: 7 ms (15.31%) | Memory: 8.4 MB (91.86%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -10,8 +10,8 @@ public:
         }
         int d=digits;
         for(int i=digits;i>0;i--){
-            long int sum1=1;
-            int digit=s%10;
+            int sum1=1;
+            long int digit=s%10;
             for(int i=d-1;i>0;i--){
                 sum1=10*sum1;
             }
